rake task bug fixes

+++ Rakefile
@@ -1,3 +1,5 @@
+require 'time'
+
 desc "create a new post"
 task :new  do |t, args|
     content_type = ENV["CONTENT_TYPE"] || 'markdown'
@@ -12,46 +14,57 @@ task :new  do |t, args|
     title.strip!
 
     now = Time.now
-    today = Date.today.to_s
-    normal = title.dup
-    normal.gsub!(/([[:space:]]|[^[:alpha:]])+/, '-')
-    normal.gsub!(/^-+/, '')
-    normal.gsub!(/-+$/, '')
-    normal.downcase!
+    filename = JekyllTask::PostTask.post_filename(now, title, content_type)
+    content = JekyllTask::PostTask.default_post_content(now, title, content_type)
 
-    create = lambda do |filename, content|
+    begin
+        file = File.open(filename, 'w')
+        file.print(content)
+        file.close
+    rescue Exception => err
+        STDERR.puts "error writing post file: #{err}"
         begin
-            file = File.open(filename, 'w')
-            file.print(content)
-            file.close
-        rescue
-            begin
-                File.delete(filename)
-            rescue Errno::ENOENT
-                return
-            rescue Exception => err
-                STDERR.puts "unable to remove post"
-                raise err
+            File.delete(filename)
+        rescue Errno::ENOENT
+        rescue Exception => err
+            STDERR.puts "error removing post file: #{err}"
+        end
+        exit 1
+    end
+
+    finalized = false
+    until finalized do
+        sh "$EDITOR #{filename}"
+        action = nil
+        while action.nil? do
+            print "done?[Yes/no/edit] "
+            resp = STDIN.readline.chomp.strip.downcase
+            resp = 'yes' if resp.empty?
+            case
+            when 'yes'.start_with?(resp)
+                action = :finalize
+            when 'no'.start_with?(resp)
+                action = :forget
+            when 'edit'.start_with?(resp)
+                action = :rework
             end
         end
-
-        until finalized do
-            sh "$EDITOR #{filename}"
-            categories = `grep '^categories:' #{filename} | awk -F: '{print $2}'`
-            categories.strip!
-            categories = categories.split
-            puts categories.inspect
-        end
+        break if action == :forget
+        finalized = action == :finalize
     end
-    filename = "_posts/#{today}-#{normal}.#{content_type}"
-    create.call(filename, (<<"HEAD"))
----
-layout: post
-title:  "#{title.gsub(/(["\\])/) do |c| "\\" + c end}"
-date:   #{now}
-categories:
----
-HEAD
+
+    if !finalized
+        begin
+            File.delete(filename)
+        rescue Errno::ENOENT
+            STDERR.puts "post file already removed"
+        rescue Exception => err
+            STDERR.puts "error removing post file: #{err}"
+        else
+            STDERR.puts "removed post file"
+        end
+        exit 1
+    end
 end
 
 task :serve
@@ -62,4 +75,29 @@ end
 task :bundle
 task :bundle do
     sh "bundle > /dev/null"
+end
+
+module JekyllTask
+    module PostTask
+        def self.post_filename(created, title, content_type)
+            date = created.to_date.to_s
+
+            normal = title.dup
+            normal.gsub!(/([[:space:]]|[^[:alpha:]])+/, '-')
+            normal.gsub!(/^-+/, '')
+            normal.gsub!(/-+$/, '')
+            normal.downcase!
+
+            "_posts/#{date}-#{normal}.#{content_type}"
+        end
+
+        def self.default_post_content(created, title, content_type)
+            content =  "---\n"
+            content << "layout: post\n"
+            content << "title:  \"#{title.gsub(/(["\\])/) do |c| "\\" + c end}\"\n"
+            content << "date:   #{created}\n"
+            content << "categories:\n"
+            content << "---\n"
+        end
+    end
 end
